requirements.txt
asgiref==3.7.2
black==23.3.0
click==8.1.3
Django==4.2.1
django-enumfields==2.1.1
djangorestframework==3.14.0
mypy-extensions==1.0.0
packaging==23.1
pathspec==0.11.1
platformdirs==3.5.1
pytz==2023.3
sqlparse==0.4.4
tomli==2.0.1
typing_extensions==4.6.2
